pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stablecoin"
version = "0.1.0"
description = 'smart contracts in a notebook'
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
keywords = []
authors = [
    { name = "Dave Bryson", email = "davebryson@users.noreply.github.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]

dependencies = [
    "matplotlib>=3.8.2",
    "simular-evm>=0.2.6",
    "mesa==2.3.0",
    "jupyter>=1.0.0",
    "seaborn",
]

[project.urls]
Documentation = "https://github.com/contracts_and_notebooks/stablecoin#readme"
Issues = "https://github.com/contracts_and_notebooks/stablecoin/issues"
Source = "https://github.com/contracts_and_notebooks/stablecoin"


[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11", "3.12"]
